Move MPI non-root event loop into a class
This helps keep things organized and moves common MPI event loop
code into a base class.
